Move existing transforms to 'single' directory
