+++ pkg/transform/single/base64Encode.go
