+++ pkg/transform/single/engine/engine_lua.go
@@ -26,7 +26,7 @@ import (
 
 	"github.com/snowplow/snowbridge/config"
 	"github.com/snowplow/snowbridge/pkg/models"
-	"github.com/snowplow/snowbridge/pkg/transform"
+	transform "github.com/snowplow/snowbridge/pkg/transform/single"
 )
 
 // LuaEngineConfig configures the Lua Engine.
