+++ pkg/transform/single/filter/snowplow_enriched_filter_context_test.go
@@ -16,7 +16,7 @@ import (
 	"testing"
 
 	"github.com/snowplow/snowbridge/pkg/models"
-	"github.com/snowplow/snowbridge/pkg/transform"
+	transform "github.com/snowplow/snowbridge/pkg/transform/single"
 	"github.com/stretchr/testify/assert"
 )
