+++ pkg/transform/single/filter/snowplow_enriched_filter_unstruct_test.go
@@ -17,7 +17,7 @@ import (
 	"testing"
 
 	"github.com/snowplow/snowbridge/pkg/models"
-	"github.com/snowplow/snowbridge/pkg/transform"
+	transform "github.com/snowplow/snowbridge/pkg/transform/single"
 	"github.com/stretchr/testify/assert"
 )
